Change dataset types in test data.
The "perma" prefixes for these dataset type names made a bit of sense
originally because they were attached to "flat" and "bias" dataset
types that didn't have a way to be associated with a validity range
but otherwise resembled the typical versions of those datasets.  But
now we're changing how validity ranges are associated with datasets,
and there will nothing to distinguish them from the usual datasets,
and hence it makes sense to drop that prefix.