Rework interfaces for database representations of topological regions.
This adds a new SpatialRegionDatabaseRepresentation class to
encapsulate how we store regions in the database, and shuffles some
code between TimespanDatabaseRepresentation and its (and the new
class's) base class, TopologicalEndpointDatabaseRepresentation.

Because we still only have one representation for spatial regions in
the database (opaque byte strings with no database-side operators),
this doesn't change functionality, but it cleans up centralizes what
was previously a lot of scattered hard-coding of how we handle regions.

It may also make it a bit easier to support database-native regions
with real operators someday in the future, but given that don't know
if we'll ever want that, that's at most a small side bonus.