Improve stringification of DataCoordinate.
In addition to adding __repr__ and one __str__ to the classes returned
by .full and .records, this makes DataCoordate.__repr__ report only
required dimensions, because that's all that's in keys().  That's good
because the inherited values() implementation returns a ValuesView with
a __repr__ that prints the __repr__ of the whole mapping, so before
this change, the keys *appearing* there would differ both from what was
actually in the ValuesView and what was in the keys (and the same was
true of items()/ItemsView, I think).