Rewrite Timespan with more careful edge-case handling and new methods.
This moves conversion to integer nanoseconds from astropy.time.Time to
Timespan, instead of at the SQLAlchemy layer, which means we can
finally make the edge-case handling (which has been improved - in that
it's now consistent - in other ways) actually the same in Python and
the DB.

This also adds methods to construct TimespanDatabaseRepresentation
instances from Timespan literals. That means we don't need to provide
methods that relate TimespanDatabaseRepresentation to Timespan
anymore.  Code that starts with a mix of those types is now
responsible for doing the conversion.

On the other hand, because Timespan and TimespanDatabaseRepresentation
can't exactly represent instants in time, we do need "overloads" for
relatinship methods that take astropy.time.Time and
sqlalchemy.sql.ColumnElement objects that represent times.